IN-172 refactor: change file structure
