+++ src/features/auth/ui/index.ts
@@ -1,5 +1,5 @@
-export * from './AgreementWithTheTerms'
-export * from './ConfirmSignUpModal'
 export * from './ExternalServicesRegistration'
-export * from './SignUpForm'
 export * from 'src/features/auth/ui/ForgotPasswordForm'
+export * from 'src/features/auth/ui/signUp/AgreementWithTheTerms'
+export * from 'src/features/auth/ui/signUp/ConfirmSignUpModal'
+export * from 'src/features/auth/ui/signUp/SignUpForm'
