+++ src/features/auth/ui/signUp/AgreementWithTheTerms.tsx
